+++ CHANGES.md
@@ -329,4 +329,4 @@
 
 ### Pre-release (work-in-progress)
 - ReSound/Beltone/Interton: fix error handling
-- Add support for Danavox XE BeMore
+- Add support for Danavox XE BeMore and Danavox Danalogic